fixed issue with too fast scrollin on large phones
When the mobile device is large, many projects fit on one page and the scroll
bar is very small. In this case the background images scroll too fast as
they depend on the size of the scrollbar. The fix is to have a larger spacer
at the end of large mobile phones in order to get a larger scrollbar.


Former-commit-id: 253d4d54ef81da419d84a1aa8ecd5b68a075ea13